Change calibration.
To be comparable with the accuracy plots provided at the end of the
last section.

Increase the depreciation rate. Lower the size of the productivity
shock and increase its persistence to match US productivity data.

The lower bound for investment is now 0.85 its steady state level (so
that we do not need to consider a crazy TFP variance to hit the lower
bound).